pyproject.toml
[project]
name = "mcp-on-cloudrun"
version = "0.1.0"
description = "MCP server on Cloud Run"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "fastmcp==2.6.1",
    "google-cloud-aiplatform>=1.71.1",
    "googlemaps>=4.10.0",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.74",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-xai>=0.2.5",
    "langgraph>=0.6.6",
    "langgraph-supervisor>=0.0.29",
    "mcp>=1.13.0",
    "mysql-connector-python>=9.4.0",
    "pydantic>=2.11.7",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "vertexai>=1.71.1",
]
